Fix for UIBarButtonItems not being styled in time
A UIBarButtonItem that is being pushed onto a navigation stack does not get styled immediately.   The reason is the UINavigationBar is always on the screen, and doesn't get the didMoveToWindow method again,  so classy has no way of knowing that these items need styling.

With this fix whenever you set the item(s) on the UINavigationItem it will do the styling immediately.