Fixes a crash issue that happened when a thread was modifying scheduledItems and another one was reading it
This problem happened when a thread was changing the scheduledItems hash table, while a timer triggered an
update in another thread. This issue was very rare but caused a crash with an exception like
'Collection <NSConcreteHashTable: 0x170134640> was mutated while being enumerated'. It was solved by guarding
the code parts that read and modify the hash table with a @synchronized block.